MozReview-Commit-ID: FA9OZyjP59N

+++ layout/style/StyleBackendType.h
@@ -12,7 +12,7 @@ namespace mozilla {
 /**
  * Enumeration that represents one of the two supported style system backends.
  */
-enum class StyleBackendType : int
+enum class StyleBackendType : uint8_t
 {
   Gecko = 1,
   Servo
